+++ ConsorcioOnline/Models/PropostaCarta.cs
@@ -7,13 +7,13 @@ namespace ConsorcioOnline.Models
     public class PropostaCarta
     {
         [Key]
-        public Int32 Id { get; set; }
+        public long Id { get; set; }
         [Required]
-        public Int32 IdCarta { get; set; }
+        public long IdCarta { get; set; }
         [Required]
-        public Int32 IdVendedor { get; set; }
+        public long IdVendedor { get; set; }
         [Required]
-        public Int32 IdComprador { get; set; }
+        public long IdComprador { get; set; }
         [Required]
         public int StatusProposta { get; set; }
         [Required]
